Sprint_5/proyect5/requirements.txt
pandas
plotly_express
streamlit
nbformat